+++ NEWS.md
@@ -2,6 +2,7 @@
 ## Minor changes
 * Exclude vignette from CRAN checks
 * add release date info to cube metadata and cube list calls
+* add auto-refresh option for sqlite tables
 
 # cansim 0.3.7
 ## Minor changes